+++ src/main/java/com/g3/elis/controller/admin/InstructorListController.java
@@ -1,6 +1,9 @@
 package com.g3.elis.controller.admin;
 
 import java.util.List;
+import java.util.Set;
+import java.util.stream.Collectors;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -18,6 +21,15 @@ public class InstructorListController {
 
     @Autowired
     UserService userService;
+    
+    @GetMapping("/admin-instructor-list")
+	public String adminInstructorList(Model model)
+	{
+		List<User> users = userService.getAllInstructors();
+	model.addAttribute("users",users);
+	model.addAttribute("content","admin/admin-instructor-list");
+	return "/admin/admin-layout";
+	}
 
     @GetMapping("/admin-instructor-list/search")
     public String searchInstructors(
