+++ src/main/java/com/g3/elis/service/UserService.java
@@ -3,7 +3,7 @@ package com.g3.elis.service;
 import java.util.List;
 
 import com.g3.elis.dto.form.UserDto;
-import com.g3.elis.model.CourseCategory;
+
 import com.g3.elis.model.User;
