+++ src/main/resources/templates/admin/admin-instructor-list.html
@@ -84,7 +84,7 @@
 									<div class="row g-4">
 
 										<!-- Card item START -->
-										<div th:each="user : ${users}" class="col-md-6 col-xxl-4">
+										<div th:each="user : ${users}"  class="col-md-6 col-xxl-4">
 											<div class="card bg-transparent border h-100">
 												<!-- Card header -->
 												<div
@@ -202,7 +202,7 @@
 											<tbody>
 
 												<!-- Table row -->
-												<tr th:each="user : ${users}">
+												<tr th:each="user : ${users}" >
 													<!-- Table data -->
 													<td>
 														<div class="d-flex align-items-center position-relative">
